\name{seraphim-package}
\alias{seraphim-package}
\alias{seraphim}
\docType{package}
\title{
A package to analyse environmental raster and spatiotemporally-referenced phylogenetic trees
}
\description{
The package "seraphim" (for "studying environmental rasters and phylogenetic informed movements"; Dellicour \emph{et al}., \emph{submitted}) can be used to (i) extract the spatio-temporal information contained in spatiotemporally-referenced trees inferred, for instance, by the continuous phylogeographic method implemented in BEAST (Lemey \emph{et al}. 2010), (ii) estimate a series of dispersion/epidemiological statistics based on these trees, (iii) compute the environmental distances associated with each phylogenetic branch for given environmental rasters, (iv) estimate the correlation between the dispersal duration and these environmental distances associated with each branch, and (v) test the statistical significance of these correlations using null models generated by randomisation procedures.
}
\author{
Simon Dellicour, Rebecca Rose, Nuno Faria, Philippe Lemey, Oliver G. Pybus
}
\references{
Dellicour S, Rose R, Pybus OG (2016). Explaining the geographic spread of emerging epidemics: a framework for comparing viral phylogenies and environmental landscape data. \emph{BMC Bioinformatics} 17: 82. \cr \cr
Dellicour S, Rose R, Faria N, Lemey P, Pybus OG (2016). SERAPHIM: studying environmental rasters and phylogenetically-informed movements. \emph{Bioinformatics} 32 (20): 3204-3206. \cr \cr
Lemey P, Rambaut A, Welch JJ, Suchard MA (2010). Phylogeography takes a relaxed random walk in continuous space and time. \emph{Molecular Biology & Evolution} 27: 1877-1885.
}
\seealso{
\url{http://beast.bio.ed.ac.uk} \cr
\url{http://www.circuitscape.org}
}
